+++ docs/conf.py
@@ -98,6 +98,8 @@ nitpicky = True
 nitpick_ignore = [
     ("py:class", "ValueT"),
     ("py:class", "MetricT"),
+    ("py:class", "InstrumentT"),
+    ("py:obj", "opentelemetry.metrics.instrument.InstrumentT"),
     # Even if wrapt is added to intersphinx_mapping, sphinx keeps failing
     # with "class reference target not found: ObjectProxy".
     ("py:class", "ObjectProxy"),